requirements_web.txt
flask
flask-cors
opencv-python
mediapipe==0.10.9
numpy
Pillow
onnxruntime
protobuf==3.20.3
